fix ssrf in og scraper, trim og-metadata surface
