+++ src/routes/api/og-image-proxy/+server.ts
@@ -1,4 +1,5 @@
 import type { RequestHandler } from './$types'
+import { isDisallowedHost } from '$lib/server/url-guard'
 
 const FETCH_TIMEOUT_MS = 8000
 const MAX_BYTES = 10 * 1024 * 1024
@@ -64,24 +65,4 @@ export const GET: RequestHandler = async ({ url, fetch: svelteFetch, setHeaders 
 	})
 
 	return new Response(upstream.body, { status: 200 })
-}
-
-function isDisallowedHost(hostname: string): boolean {
-	const h = hostname.toLowerCase()
-	if (h === 'localhost' || h === '' || h === '::1') return true
-	// Literal IPv4 checks — block loopback, private ranges, link-local, reserved.
-	const v4 = h.match(/^(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})$/)
-	if (v4) {
-		const [a, b] = [Number(v4[1]), Number(v4[2])]
-		if (a === 10 || a === 127) return true
-		if (a === 169 && b === 254) return true
-		if (a === 172 && b >= 16 && b <= 31) return true
-		if (a === 192 && b === 168) return true
-		if (a === 0) return true
-	}
-	// IPv6 literals in URLs are bracketed; URL.hostname strips brackets.
-	if (h.includes(':') && (h.startsWith('fc') || h.startsWith('fd') || h.startsWith('fe80'))) {
-		return true
-	}
-	return false
 }
